Move profile related files under profile directory
